+++ testing/test_sample.py
@@ -4,6 +4,9 @@ import inspect as ins
 from typing import get_args
 from pytest import fixture
 import pytest
+import matplotlib.pyplot as plt
+import matplotlib as mpl
+from itertools import repeat
 
 cali_sample_names = [['ExpDat_250605_dd_Calciumoxalat_02_spline',
  'ExpDat_250605_dd_Calciumoxalat_03_spline',
@@ -24,14 +27,14 @@ def get_cali_worklist(request):
 @fixture(params = [(name, profile) for name, profile in zip(sample_names, profiles)], ids=profiles)
 def get_sample(request):
     name, profile = request.param
-    print(name, profile)
     yield Sample(name, profile=profile)
 
-@fixture(params = ((name, profile) for name, profile in zip([sample_names]*3, profiles*3)), ids=profiles*3)
+@fixture(params = ((samplename, profile) for samplename, profile in zip(sample_names*3, profiles*3)), ids=profiles*3)
 def get_worklist(request):
     def init_options():
-        for fun in [lambda x, y: x, lambda x, y: Sample(x[0], profile=y), lambda x, y: [Sample(s, profile=y) for s in x]]:
-            init = fun(request.param)
+        yield Worklist(request.param[0], profile = request.param[1])
+        for fun in [lambda x, y: Sample(x, profile=y), lambda x, y: [Sample(x, profile=y)]]:
+            init = fun(*request.param)
             yield Worklist(init)
     return init_options
 
@@ -55,27 +58,31 @@ class TestSampleInit:
 @pytest.mark.usefixtures("get_sample")
 class TestSamplePlot:
     def test_plot_sample(self, get_sample):
-        plot_opts = get_args(ins.signature(get_sample.plot).parameters["plot"])
-        for plot in plot_opts:
-            get_sample.plot(plot)
+        plot_params = ins.signature(get_sample.plot).parameters["plot"].default
+        plot_opts = get_args(plot_params)
+        assert all(isinstance(get_sample.plot(plot), mpl.axes.Axes) for plot in plot_opts) 
 
 @pytest.mark.usefixtures("get_cali_worklist", "get_sample")
 #@pytest.mark.parametrize("get_sample, get_cali_worklist", [((sample_names[0], "Netzsch"), cali_sample_names)], indirect=True)
 class TestSampleCali:    
     def test_calibration(self, get_sample, get_cali_worklist):
+        if get_sample.profile != get_cali_worklist.profile:
+            pytest.skip("Different profiles")
+        
         sample = get_sample
         wl = get_cali_worklist
-        sample.calibrate(worklist=wl, mode="recalibrate", molecular_formulas = {'QMID(s:1|m:18)/A': 'H2O', 'QMID(s:2|m:44)/A': 'CO2'})
+        sample.calibrate(worklist=wl, mode="recalibrate", molecular_formulas = {'QMID(s:1|m:18)/A': 'H2O', 'QMID(s:2|m:44)/A': 'CO2'},plot=True)
         sample = get_sample
         assert isinstance(get_sample.stats , pd.DataFrame)
         assert isinstance(get_sample.linreg , pd.DataFrame)
         assert isinstance(get_sample.xcali , pd.DataFrame)
         assert isinstance(get_sample.ycali , pd.DataFrame)
 
-@pytest.mark.usefixtures("get_cali_worklist", "get_sample", "get_worklist")
+@pytest.mark.usefixtures("get_worklist")
 class TestWorklistInit:
     def test_init_worklist(self, get_worklist):
         for worklist in get_worklist():
+            print(worklist)
             assert isinstance(worklist.samples, list)
             assert isinstance(worklist.name, str)